resfactor: разбиение на компоненты
