+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,7 @@
+import './Footer.css';
+
+function Footer() {
+  return <footer className="Footer"></footer>;
+}
+
+export default Footer;
