+++ src/components/Header/Header.jsx
@@ -1,0 +1,7 @@
+import './Header.css';
+
+function Header() {
+  return <header className="Header"></header>;
+}
+
+export default Header;
